+++ machines/hostup1/configuration.nix
@@ -21,12 +21,8 @@
 
   # hostup1 has 4 cores.
   # Build four jobs at a time using maximum of 1 cores per job.
-  nix.settings.max-jobs = 4;
-  nix.settings.cores = 1;
-
-  nix.settings.trusted-users = [
-    "daniel"
-  ];
+  nix.settings.max-jobs = lib.mkForce 4;
+  nix.settings.cores = lib.mkForce 1;
 
   # Enable the OpenSSH daemon.
   services.openssh.enable = true;